Some organisation and configuration
Added an organiser for the saving of the configuration. It will save
the list of recipients followed by the list of places it was already
sent (or is being told not to end). Also changed the way that the
recipients pane updates. Now it will update the recipients list after
sending.