Major bug fixes
Fixed something which fucked up loading (it was in the fileReader,
asking for things before they were loaded). Fixed the saving system
attempting to append ‘.txt’ to everything (it was using toPath()
instead of getName()). Fixed Marconi closing all windows all the time
(added a Thread.sleep in a new output method).