Added system to account for preexisting files.
The system will revise and rewrite the entire file such that corruption
of the file system is avoided. There is also code on the subject of the
new Apache Commons CLI and Lang libraries which will be introduced (I’m
tired of writing it all. Better to gaze on the shoulders of giants.)