update to alpha version 6
Started and finished prototype of the recruiter. a number of issues
obviously still exist, but they are being resolved.

1. communique parser now uses the hashset algorithm for O(n) time
2. CReader now supports attempts to read the old Communique files
3. ‘txt’ extension now guaranteed, new accelerator keystrokes,
recruiter caller now built. added system via flag:recruit.

Release of Communique 6 upcoming when features in the programming of
the new system are finally implemented and recruiter is available for
Marconi.