lots of bugs
the filtering mechanism failed. it was fixed. a filter was added to the
output for marconi’s sentList, which now gets rid of duplicates and
deals with some systematic errors from the early days. The JInfoFetcher
is now a singleton object. MarconiRecruiter now directly updates its
own dynamic copy of the sentList, which it did not in the past (because
I fucked up).