Some changes, bug fixes
The program now has full integration between the Recruiter exclusion
system and the manual version of implementation. Also, some changes to
Streams basically for self-documentation.

There is a rare bug if NationStates doesn’t populate data in the
recruiter. That’s been solved with better error-handling.

This version has been used as the basis for the 6.3 release.