Fixed bug where recipients can pop up twice
Solved by switching from list to linkedhashset, which is how this issue
was solved in the original parser. Chose this solution over an active
duplicate checking solution because it would be faster this way. Also
worked on filtering code, it is now faster and doesn’t reparse parsed
data.

Added some tooltips to the program, changed acknowledgements to reflect
bug report source. Thank you, Wallenburg. Also added hashCode and
equals methods to CommuniqueRecipient.