various changes, see comments
it’s been so long since my last commit that i don’t really recall
exactly what it is that i’ve done in full exactitude.

there have been changes to CommuniqueConfig, which has been rewritten
to make programmatic changes much more easy, along with a bug-fix to an
issue where each load of the configuration file would make prepend
‘nation:’ to whatever the element was. There have been additions to
make it update old files to the current mechanism. Here, the
recipients-sentList distinction has been deprecated. Everything is now
kept in a single list of recipients.

the philosophy on how it operates now is that the actual list itself is
private. what is shown and given to the program is a created list of
communique recipients, which can be edited as necessary.

some work on the token translator was done, along with the parser, to
make it account for different conditions.

Communique, as a GUI program, now logs console data to a newly created
log file in all situations… because most WIndows computers are set up
stupid, meaning you cannot simply `java -jar` the program.

Marconi now uses a file lock to prevent multiple instances of Marconi
from running. Communique does not have this functionality because it is
a GUI program, and you should be able to figure that out yourself.

both marconi and communique have stopped using a direct delay for
recruiting. they now delay on a timer which checks every second until
it the delay is completed. around second 170, the program queries for
the next recipient, shortening the delay in-between telegrams.
javatelegram also implements a similar system for campaigns, which
skips the recipient if it is found not to be a campaign telegram, but
doing so within the delay time.

various additions to the documentation have been affected. most
everything has stopped using `System.err…`, and now uses the
`java.util.logging` solution.