My NS API and JavaTelegram packages have been subsumed into Communique. Many changes emerge from refactoring needed to deal with that.
Formatting has changed in some places.

Searching by region tags has been implemented with `region_tag:*` that
is decomposed to regions and then to nations within those regions. It
is unclear whether there is a faster way to do this, given need to poll
every single region at least once. Use of daily dump could be done, but
would require download of a large file.

Search by region tag also necessitates end of auto-parse for single
liners with commas. This was automatically done in the past, but now, a
line like `region:Europe, +tag:wa` will no longer be separated auto-
matically.

Version bumped to version 11 because of changes to file structure that
are not backwards compatible.