started daemon mode coding for real
I realised that the problem wasn't at Communique's level, but rather, at the lowest level itself; the `rewrite_javatelegram` branch is meant to resolve the fundamental issue associated with the old library: its rigidity in demanding that all the recipients be enumerated before doing anything.

In that, the first thing to do is decouple the recipients from any fully enumerated list; this is where the idea of a monitor comes up. A monitor would generate a series of recipients.

To handle future recipients, the client itself would need also to process them in a buffer, using a queue and a feed limit to keep the buffer fresh, because the buffer can be discarded as it updates. Also want to keep the extensibility and auto-filtering of the earlier version, which requires providing some ability to apply functions to the existing data.