fix communiqueeditor -> config -> parsed recipients desync bug
updates to the communique editor change the pre-parse recipients. but after they are turned into parsed recipients they are shunted to a config object which isn't itself updated. after that config object is placed in the monitor it doesn't get updated which means that it keeps parsing against the same list of original recipients.

this changes that into an interface which allows the editor, implementing getConfig to be passed; this doesn't break marconi code much because marconi is its own implementation.

this whole thing emerges because CommuniqueEditor constructs the config when requested rather than automatically updating the config on every tiny change which is what the old in-Communique editors did. it also seems to sit uneasily with how CommuniqueEditor basically pretends that the config is immutable and how the other code assumes its mutability. either way, some tensions can be fixed by throwing in mapping code.